[html] Update tag arrays to HTML5 spec
Updated the tag arrays in html2json to the HTML5 spec referenced in the
htmlparser file. Also removed the <br> tag from the inline array as it
was causing issues with text fields where referenced.
e.g.

<th>
"Hello"
<br>
"World"
</th>
<th>My name is
<strong>Slim Shady</strong>
</th>

Because the htmlparser would treat the <br> tag as inline, it was
expecting a closing tag, but that never happened. The "Hello World" text
would be squashed and included into the Slim Shady text because the
<strong> tag would be the next inline tag recognized.
This would cause the "Hello World" text to be associated with the second
<th> tag rather than the first.
Instead now the <br> tags are treated as empty and and the "Hello World"
will be included as text in the first <th> tag.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>